requirements.txt
flask~=1.1.0
flask-restplus~=0.13.0
flask-sqlalchemy~=2.4.0
flask-httpauth~=3.3.0
